Finish search and remove WebAppTest
